feat: 이벤트 목록 조회 구현 (#17)
* 🔒️ JWT 관련 설정 리펙토링
* ♻️ 회원 간편로그인 리펙토링 (JWT 수정사항 반영)
* ♻️ 컨트롤러 수정 (Pageable 적용)
* ✨ 이벤트 목록 조회 구현
* 💚 PR 테스트 워크플로우 오류 수정

+++ src/main/java/hana/account/controller/TransactionController.java
@@ -70,7 +70,7 @@ public class TransactionController {
             @RequestParam("end") String end,
             @RequestParam("type") String type,
             @RequestParam("sort") String sort,
-            @RequestParam("page") Long page) {
+            @RequestParam("page") Integer page) {
         return null;
     }
